+++ subsys/sd/sd_utils.h
@@ -28,6 +28,15 @@ enum sd_return_codes {
 	SD_RESTART = 3,
 };
 
+/* Checks SD status return codes */
+static inline int sd_check_response(struct sdhc_command *cmd)
+{
+	if (cmd->response_type == SD_RSP_TYPE_R1) {
+		return (cmd->response[0U] & SD_R1_ERR_FLAGS);
+	}
+	return 0;
+}
+
 /* Delay function for SD subsystem */
 static inline void sd_delay(unsigned int millis)
 {
